requirements.txt
lxml
natsort
gensim
keras
tensorflow
numpy
javaobj-py3
